Hike/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
pandas==2.1.4
python-dotenv==1.0.0

# Social media APIs
python-twitter-v2==0.7.8
linkedin-api==2.0.0
selenium==4.15.2

# AI and NLP
openai==1.3.7
spacy==3.7.2
nltk==3.8.1

# Database and storage
airtable-python-wrapper==0.15.3
notion-client==2.2.1
redis==5.0.1

# Notifications
slack-sdk==3.26.1
boto3==1.34.0

# Utilities
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
lxml==4.9.3
html5lib==1.1

# Web Interface
flask==3.0.0
flask-socketio==5.3.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1 